Implement code minification
Code can now be minified / uglified without breaking. It was a matter of specifying explicit dependency injection for all controller, service, config and run methods. Previously this was missing from config and run. 'ng-strict-di' attribute was added to the body tag of index.html in order to force this good coding habbit.